Require *.svg files fix
