+++ packages/ckeditor5-ui/package.json
@@ -9,7 +9,7 @@
     "ckeditor5-lib",
     "ckeditor5-dll"
   ],
-  "main": "src/index.js",
+  "main": "src/index.ts",
   "dependencies": {
     "@ckeditor/ckeditor5-utils": "^35.1.0",
     "@ckeditor/ckeditor5-core": "^35.1.0",
