--- packages/ckeditor5-ui/src/custom.d.ts
@@ -1,9 +1,0 @@
-/**
- * @license Copyright (c) 2003-2022, CKSource Holding sp. z o.o. All rights reserved.
- * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
- */
-
-declare module '*.svg' {
-	const content: string;
-	export default content;
-}
